requirements.txt
dill>=0.3.6
easydict>=1.10
freezegun>=1.2.2
hypernetx>=1.2.5
imageio>=2.31.1
joblib>=1.3.1
kaleido>=0.1.0.post1
matplotlib>=3.7.2
molsets>=0.3.1
networkx>=2.8.8
numpy>=1.24.4
pandas>=2.0.3
plotly>=5.15.0
pyemd>=1.0.0
pytest>=7.4.0
pytz>=2023.3
PyYAML>=6.0.1
rdkit>=2023.3.2
scikit_learn>=1.3.0
scipy>=1.11.1
TopoNetX>=0.0.2
torch>=2.0.1
tqdm>=4.65.0
~aleido>=0.2.1
